Added functionality. DB now properly connected
Added login functionality, including checks for incorrect credentials and whether or not the account has been banned. Added code to include the super-global variable $_SESSION for keeping track of login status between pages and to remember the last visited page so the user is returned there upon login (currently not very useful, but this will enable people to quickly login if they found a server they wish to apply for whitelist for and be returned there directly after a successful login)

Also uploaded the current (unfinished but currently functional) database, safed in the database folder.